feat : H2 DB에 생성할 Question 테이블 - 객체 맵핑
* questionId 컬럼 : Primary Key

* toStringDateTime() 메소드를 제거하고 getCreatedDateTimetoString()를 새로 정의. 게시판에서 호출되는 날짜 데이터에 대한 getter 메소드로 지정
