+++ src/main/java/com/codesquad/qna/domain/Question.java
@@ -1,0 +1,73 @@
+package com.codesquad.qna.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+@Entity
+public class Question {
+    @Id
+    @GeneratedValue
+    private long questionId;
+    @Column(nullable = false)
+    private String writer;
+    @Column(nullable = false)
+    private String title;
+    private String contents;
+    @Column(nullable = false)
+    private LocalDateTime createdDateTime;
+
+    public void setQuestionId(long questionId) {
+        this.questionId = questionId;
+    }
+
+    public void setWriter(String writer) {
+        this.writer = writer;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public void setContents(String contents) {
+        this.contents = contents;
+    }
+
+    public void setCreatedDateTime(LocalDateTime createdDateTime) {
+        this.createdDateTime = createdDateTime;
+    }
+
+    public long getQuestionId() {
+        return questionId;
+    }
+
+    public String getWriter() {
+        return writer;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getContents() {
+        return contents;
+    }
+
+    public String getCreatedDateTimetoString() {
+        return DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").format(createdDateTime);
+    }
+
+    @Override
+    public String toString() {
+        return "Question{" +
+                "questionId=" + questionId +
+                ", writer='" + writer + '\'' +
+                ", title='" + title + '\'' +
+                ", contents='" + contents + '\'' +
+                ", createdDateTime=" + getCreatedDateTimetoString() +
+                '}';
+    }
+}
